.plot_creation

Function to create the main heatmap plot

Usage

.plot_creation(
  PPV_NPV = "PPV",
  one_out_of = TRUE,
  min_Prevalence,
  max_Prevalence,
  min_FP = 0,
  max_FP,
  max_FN,
  min_FN,
  PPV_melted,
  steps_matrix = 100,
  decimals_x,
  decimals_y,
  label_title = "",
  label_subtitle = "",
  translated_labels = translated_labels
)

Arguments

PPV_NPV

Should show PPV or NPV ["PPV", "NPV"]

one_out_of

Show y scale as 1 out of x [TRUE, FALSE] FALSE by default

min_Prevalence

[x] out of y prevalence of disease: [1-Inf]

max_Prevalence

x out of [y] prevalence of disease: [1-Inf]

min_FP

Minimum False Positive rate: [0-100]

max_FP

Maximum False Positive rate: [0-100]

max_FN

Maximum False Negative rate: [0-100]

min_FN

Minimum False Negative rate: [0-100]

PPV_melted

DF out of .createPPVmatrix()

steps_matrix

width of PPV/NPV matrix. 100 by default

decimals_x

Number of decimals to show in x axis [0-2]

decimals_y

Number of decimals to show in y axis [0-2]

label_title

Title for the plot

label_subtitle

Subtitle for the plot

translated_labels

Translated labels for plot. Output of .translate_labels()