Show minimum possible prevalence given the test characteristics

Given a FP and a desired PPV, what is the Minimum Prevalence of a Condition

min_possible_prevalence(Sensitivity = 95, FP_test = 1, min_PPV_desired = 90)

Arguments

Sensitivity
Sensitivity of the test: [0-100]

FP_test
False positive rate (1-Specificity): [0-100]

min_PPV_desired
Which PPV is what you consider the minimum to trust a positive result in the test: [0-100]

Value

A description showing the minimum necessary prevalence.

Examples

# Example 1
min_possible_prevalence(Sensitivity = 99.9, FP_test = .1, min_PPV_desired = 70)

#> To reach a PPV of 70% when using a test with 99.9% Sensitivity and 0.1% False Positive Rate, you need a prevalence of at least 1 out of 429

"To reach a PPV of 70 when using a test with 99.9 % Sensitivity and 0.1 % False Positive Rate, 
you need a prevalence of at least 1 out of 429"

#> [1] "To reach a PPV of 70 when using a test with 99.9 % Sensitivity and 0.1 % False Positive Rate, \nyou need a prevalence of at least 1 out of 429"

# Example 2
min_possible_prevalence(100, 0.1, 98)

#> To reach a PPV of 98% when using a test with 100% Sensitivity and 0.1% False Positive Rate, you need a prevalence of at least 1 out of 21

"To reach a PPV of 98 when using a test with 100 % Sensitivity and 0.1 % False Positive Rate, 
you need a prevalence of at least 1 out of 21"

#> [1] "To reach a PPV of 98 when using a test with 100 % Sensitivity and 0.1 % False Positive Rate, \nyou need a prevalence of at least 1 out of 21"